feat(poly): graph-evaluator impl
